Refactor scales loadbalancer to be robust from ZK failures
Previously (and incorrectly), the ZK connection and ServerSet
are initialized when the LoadBalancerSink is constructed.
This change moves the connection establishment and bootstrapping
to the Open() method, where it occurs asynchronously.  Incoming
messages block until the sink has opened at least one node,
(or there are no nodes to open).  Messages waiting for too long
will be timed out normally via the ClientTimeoutSink.

This refactor also correctly cleans up ServerSet background
worker greenlets, which previously were leaked when a LoadBalancer
was shut down.